be/requirements.txt
qiskit
psutil
qiskit-aer
mqt.qcec
quokka_sharp
python-dotenv
